src/app/requirements.txt
streamlit>=1.38.0
pydeck>=0.9.0
databricks-sdk>=0.30.0
databricks-sql-connector>=3.0.0
srai[torch]>=0.8.0
geopandas>=0.14.0
shapely>=2.0.0
h3>=4.0.0
scikit-learn>=1.3.0
geopy>=2.4.0
numpy>=1.24.0
pandas>=2.0.0
